fix(tui): code-review sweep — sentinels, surface mirror tests, termsim coverage + docs
Apply a batch of disjoint, confirmed review findings:

1. TranscriptHome "scroll to top" used `u16::MAX as usize` (capped at
   65535), defeating the usize widening; use `usize::MAX` to match the
   `TRANSCRIPT_OVERLAY_SCROLL_BOTTOM` precedent. Updated the Home test.

2. `wrap_entries_lines_match_overlay_rows` only covered `expand_all = true`.
   Loop over `[true, false]`, seeding a long (foldable) tool result and a
   coalesced tool run so the collapsed branch is byte-identical to
   `transcript_lines_for_overlay` too.

3. Rename the misleading `RowCacheKey.subagent_active` flag-named field to
   `subagent_source_hash` (it holds a discriminator hash, not a bool).

4. Add an emission-path test for the VS Code wide-moon fix: a moon glyph at
   the last content column of an N-wide buffer is dropped by
   `emit_buffer_row_styled` (via `term_display_width`) so it can't overflow
   the row; a control case proves it emits with room to spare.

5. termsim coverage (term-matrix feature): split_frames clamping
   (offset-beyond-len, out-of-order/empty marks, empty buffer); the
   frameless-log fallback in BOTH backends (vt100 leaves the grid blank,
   alacritty replays the whole stream); and a CJK wide-glyph survives-reflow
   test for the alacritty leg.

6. Fix stale "scaffold/stub/empty step lists" docs that no longer match the
   implemented code (termsim.rs, emulator.rs, scenario.rs, the two backend
   test helpers, and xtermcheck/replay.js's translateToString trimRight note).

7. Narrow the transcript_surface module doc: the row model currently mirrors
   the OVERLAY surface (`transcript_lines_for_overlay`) only; the main view
   (`transcript_lines_for_render`) diverges, and wiring it (Phase 4) needs
   wrap_entries parameterized by surface.

+++ crates/squeezy-tui/src/lib.rs
@@ -3285,7 +3285,7 @@ fn dispatch_keymap_action(app: &mut TuiApp, agent: &mut Agent, key: KeyEvent) ->
                 return false;
             }
             if app.input.is_empty() {
-                set_active_transcript_scroll_from_bottom(app, u16::MAX as usize);
+                set_active_transcript_scroll_from_bottom(app, usize::MAX);
                 true
             } else {
                 false
